Numbered companies
This is an updated CTE list where I numbered every individual company on the dataset. Since several companies appear on the list twice, it was important to make sure that there is not an inflation of the numbers, so this CTE was used to number each of them. This will likely be altered if some numbers did not populate.